Quitar getters y setters
