+++ backend/libs/entities/src/classes/customer.entity.ts
@@ -1,7 +1,7 @@
 
 import { Column, Entity, OneToMany, OneToOne, JoinColumn } from "typeorm";
 import Generic from "./generic.entity";
-import Address from "./adress.entity";
+import Address from "./address.entity";
 import Order from "./order.entity";
 
 @Entity()
@@ -21,44 +21,4 @@ export default class Customer extends Generic {
 
     @OneToMany(() => Order, order => order.customer, { nullable: true })
     orders: Order[];
-
-    setName(name: string): void {
-        this.name = name;
-    }
-
-    setLastName(lastName: string): void {
-        this.lastName = lastName;
-    }
-
-    setPhoneNumber(phoneNumber: string): void {
-        this.phoneNumber = phoneNumber;
-    }
-
-    setAddresses(address: Address): void {
-        this.addresses = address;
-    }
-
-    setCustomerOrders(orders: Order[]): void {
-        this.orders = orders;
-    }
-
-    getName(): string {
-        return this.name;
-    }
-
-    getLastName(): string {
-        return this.lastName;
-    }
-
-    getPhoneNumber(): string {
-        return this.phoneNumber;
-    }
-
-    getAddresses(): Address {
-        return this.addresses;
-    }
-
-    getCustomerOrders(): Order[] {
-        return this.orders;
-    }
 }
